// web/src/api/generated/models/CapabilitiesView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CapabilitiesView
 */
export interface CapabilitiesView {
    /**
     * 
     * @type {number}
     * @memberof CapabilitiesView
     */
    expiresInSeconds: number;
    /**
     * 
     * @type {boolean}
     * @memberof CapabilitiesView
     */
    localPassword: boolean;
    /**
     * 
     * @type {Array<string>}
     * @memberof CapabilitiesView
     */
    oidcConnections: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof CapabilitiesView
     */
    passkey: boolean;
}

/**
 * Check if a given object implements the CapabilitiesView interface.
 */
export function instanceOfCapabilitiesView(value: object): value is CapabilitiesView {
    if (!('expiresInSeconds' in value) || value['expiresInSeconds'] === undefined) return false;
    if (!('localPassword' in value) || value['localPassword'] === undefined) return false;
    if (!('oidcConnections' in value) || value['oidcConnections'] === undefined) return false;
    if (!('passkey' in value) || value['passkey'] === undefined) return false;
    return true;
}

export function CapabilitiesViewFromJSON(json: any): CapabilitiesView {
    return CapabilitiesViewFromJSONTyped(json, false);
}

export function CapabilitiesViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): CapabilitiesView {
    if (json == null) {
        return json;
    }
    return {
        
        'expiresInSeconds': json['expiresInSeconds'],
        'localPassword': json['localPassword'],
        'oidcConnections': json['oidcConnections'],
        'passkey': json['passkey'],
    };
}

export function CapabilitiesViewToJSON(json: any): CapabilitiesView {
    return CapabilitiesViewToJSONTyped(json, false);
}

export function CapabilitiesViewToJSONTyped(value?: CapabilitiesView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'expiresInSeconds': value['expiresInSeconds'],
        'localPassword': value['localPassword'],
        'oidcConnections': value['oidcConnections'],
        'passkey': value['passkey'],
    };
}