// web/src/api/generated/models/RecentAuthenticationClient.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const RecentAuthenticationClient = {
    web: 'web',
    android: 'android'
} as const;
export type RecentAuthenticationClient = typeof RecentAuthenticationClient[keyof typeof RecentAuthenticationClient];


export function instanceOfRecentAuthenticationClient(value: any): boolean {
    for (const key in RecentAuthenticationClient) {
        if (Object.prototype.hasOwnProperty.call(RecentAuthenticationClient, key)) {
            if (RecentAuthenticationClient[key as keyof typeof RecentAuthenticationClient] === value) {
                return true;
            }
        }
    }
    return false;
}

export function RecentAuthenticationClientFromJSON(json: any): RecentAuthenticationClient {
    return RecentAuthenticationClientFromJSONTyped(json, false);
}

export function RecentAuthenticationClientFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecentAuthenticationClient {
    return json as RecentAuthenticationClient;
}

export function RecentAuthenticationClientToJSON(value?: RecentAuthenticationClient | null): any {
    return value as any;
}

export function RecentAuthenticationClientToJSONTyped(value: any, ignoreDiscriminator: boolean): RecentAuthenticationClient {
    return value as RecentAuthenticationClient;
}